+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Missing part level bar segments and "glitchy" rendering when using HD44780 LCDs.
+- Missing part level bar "bases" when using HD44780 LCDs.
+
 ## [0.8.1] - 2021-01-03
 
 ### Fixed